[Avatar]: fix review comments (tests and naming)
