+++ uui-components/src/widgets/Avatar.tsx
@@ -20,11 +20,11 @@ export interface AvatarProps extends IHasCX, IHasRawProps<React.ImgHTMLAttribute
 }
 
 const AvatarComponent = (props: AvatarProps, ref: React.ForwardedRef<HTMLImageElement>) => {
-    const [errorMode, setErrorMode] = React.useState<boolean>(false);
+    const [isError, setIsError] = React.useState<boolean>(false);
 
     function onError() {
-      if (!errorMode) {
-        setErrorMode(true);
+      if (!isError) {
+        setIsError(true);
       }
     }
     return (
@@ -34,7 +34,7 @@ const AvatarComponent = (props: AvatarProps, ref: React.ForwardedRef<HTMLImageEl
                 className={ cx(css.avatar, props.cx) }
                 width={ props.size }
                 height={ props.size }
-                src={ (props.isLoading || !props.img || errorMode) ? 'https://static.cdn.epam.com/uploads/690afa39a93c88c4dd13758fe1d869d5/EPM-UUI/Images/avatar_placeholder.jpg' : props.img }
+                src={ (props.isLoading || !props.img || isError) ? 'https://static.cdn.epam.com/uploads/690afa39a93c88c4dd13758fe1d869d5/EPM-UUI/Images/avatar_placeholder.jpg' : props.img }
                 alt={ props.alt }
                 onError={ onError }
                 { ...props.rawProps }
